Refactor book routes to nested index
